Fix race condition in library loader
This race would, on second launch of the game (after installing BSIPA)
cause Ionic.Zip to not be loaded properly. This happened because the
load would happen concurrently with a load on another thread, which led
ultimately happened before the scanner reached Ionic.Zip in the
libraries folder. Thus, when the loader checked for it, it was not
present in the dictionary.